requirements.txt
discord.py>=2.3.2
apscheduler>=3.10.1
pytz